+++ src/test/java/com/github/katanikotael/jrtb/command/CommandContainerTest.java
@@ -1,0 +1,44 @@
+package com.github.katanikotael.jrtb.command;
+
+import com.github.katanikotael.jrtb.services.SendBotMessageService;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
+
+import java.util.Arrays;
+
+@DisplayName("Unit-level testing for CommandContainer")
+class CommandContainerTest {
+
+    private CommandContainer commandContainer;
+
+    @BeforeEach
+    void setUp() {
+        SendBotMessageService sendBotMessageService = Mockito.mock(SendBotMessageService.class);
+        commandContainer = new CommandContainer(sendBotMessageService);
+    }
+
+    @Test
+    void shouldGetAllTheExistingCommands() {
+        //when-then
+        Arrays.stream(CommandName.values())
+                .forEach(CommandName -> {
+                    Command command = commandContainer.retrieveCommand(CommandName.getCommandName());
+                    Assertions.assertNotEquals(UnknownCommand.class, command.getClass());
+                });
+    }
+
+    @Test
+    void shouldReturnUnknownCommand() {
+        //given
+        String unknownCommand = "/sdfgsd";
+
+        //when
+        Command command = commandContainer.retrieveCommand(unknownCommand);
+
+        //then
+        Assertions.assertEquals(UnknownCommand.class, command.getClass());
+    }
+}